+++ packages/plugin-node-modules/sources/NodeModulesLinker.ts
@@ -167,10 +167,10 @@ class NodeModulesInstaller implements Installer {
 
     for (const [descriptor, locator] of dependencies) {
       const target = !structUtils.areIdentsEqual(descriptor, locator)
-        ? [structUtils.requirableIdent(locator), locator.reference] as [string, string]
+        ? [structUtils.stringifyIdent(locator), locator.reference] as [string, string]
         : locator.reference;
 
-      slot.pnpNode.packageDependencies.set(structUtils.requirableIdent(descriptor), target);
+      slot.pnpNode.packageDependencies.set(structUtils.stringifyIdent(descriptor), target);
     }
   }